% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_column_names.R
\name{format_column_names}
\alias{format_column_names}
\title{format_column_names}
\usage{
format_column_names(input_data_frame)
}
\arguments{
\item{input_data_frame}{a data frame or data table object}
}
\value{
A data frame or data table object
}
\description{
Formats column names in data frame in a uniform style
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
format_column_names
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
