% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_cibersortx.R
\name{post_process_cibersortx}
\alias{post_process_cibersortx}
\title{post_process_cibersortx}
\usage{
post_process_cibersortx(input_file, output_dir)
}
\arguments{
\item{input_file}{Single file name.}

\item{output_dir}{Path to the output folder.}
}
\value{
A path to the rds file.
}
\description{
Formats CIBERSORTx output file
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
post_process_cibersortx
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
